Fix lots of warnings.
Unlike previous warning-fixes, these are entangled and overlapping
so I've regretfully bundled them into a catch-all commit. There are
a few main themes;
- unused code
- unused variables
- unused parameters, for these I have avoided using attributes out
  of fear of the number of platforms/toolchains that might break,
  so I've used a corny "absorb_unused" trick. ("Add" all unused
  parameters to a static global. The compilers don't seem to notice
  or care that the resulting global goes unused...)
- const issues. To resolve const/non-const warnings, I set off the
  usual constification chain-reaction and followed it to its
  logical conclusion. I didn't const anything else, but saw plenty
  of similar-looking things that could be constified.
- implicit function declarations. I've added header declarations,
  and in a couple of places adding missing headers.
- signed/unsigned comparisons. I've casted where appropriate.
- "shift of negative value". I've forced some MASK defines to be
  unsigned to avoid this kind of compiler warning.
- removed some unused defines.

+++ Src/Ant32/Lib32/ant32_expand_op.c
@@ -575,7 +575,7 @@ int asm_expand_op(ant_op_t op_code, ant_asm_stmnt_t *stmnt, char *b_memory,
     return (1);
   }
 
-  if (remaining < pseudo_op[op_id].size) {
+  if (remaining < (unsigned int)pseudo_op[op_id].size) {
     sprintf(AntErrorStr, "instruction memory overflow");
     return (1);
   }
